kbuild: treat char as always unsigned
The C standard makes it implementation defined whether a plain char is
unsigned or signed and the architectures where barebox is compiled for
differ in that, e.g. chars are traditionally unsigned on ARM, but on x86
for example they tend to be signed.

This caused different bugs[1][2][3] in the past, especially around
behavior when casted to int. Let's instruct the compiler to treat char
as always unsigned. This may fix some issues that flew under the radar
so far, but also break drivers that were compiled and used only for
specific architectures, which implicitly assumed char is signed, which
we'll have to fix.

Linux is also being compiled with the same flag since 2022 with Linux
commit 3bc753c06dd0 ("kbuild: treat char as always unsigned").

[1]: 02a3c4f39690 ("readkey: keys are unsigned char")
[2]: f8fccf2ef8c9 ("mci: fix wrong sd/mmc/emmc card size computation for arch where char is signed")
[3]: https://lore.barebox.org/barebox/20250422053641.3435585-1-a.fatoum@pengutronix.de/T/#u

Cc: Yann Sionneau <ysionneau@kalrayinc.com>
Cc: Michael Olbrich <mol@pengutronix.de>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250422063910.126829-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>