virtio: make more robust in face of misbehaving devices
The virtio device drivers assume at a number of places that polling
indefinitely with virtqueue_get_buf will eventually succeed.

It was observed that sometimes after a warm reboot inside QEMU, barebox
finds itself inside an infinite poll loop.

This needs to be further analyzed, of course, but it's already and
improvement for barebox to recover from such situations by returning an
error code or at least printing an error message.

It's unlikely though that the system will be completely functional
after this, but it gives the user a fighting chance.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250422075658.221273-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>