fastboot: only populate variables explicitly asked for
To allow flashing an image with a partition table followed by an image
for a specific partition, the fastboot implementation was changed to
evaluate partitions on demand.

As there are partition-specific variables and the magic "all" variable,
which lists all variables, all partitions are evaluated in getvar
unconditionally, when only a variable for a specific partition is needed.

When one partition doesn't exist and is not marked as optional, this
leads to an error, even if something innocuous as the general
max-download-size is requested.

Change this, so only partitions that were explicitly asked for have their
variables populated on demand, unless "all" variables were requested.

Fixes: 6a191155be4e ("fastboot: evaluate fastboot partitions when used")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250428125217.1931365-5-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>