fastboot: check for file existence before flashing/erasing
We used to have an init time check for whether exported files exist or
not, but this went away and was replaced by a check in getvar.

If a client omits getvar and directly calls flash for a non-existent
file, barebox will create it automatically, thereby turning every
partition as if it had the create flag set.

Add an explicit check for whether a file exists into flash/erase
and proceed only if it does or the create flag is set.

The optional flag needs not to be checked here as it's only relevant
when enumerating available partitions.

Fixes: 6a191155be4e ("fastboot: evaluate fastboot partitions when used")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250428125217.1931365-7-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>