clocksource: timer-ti-dm: adapt omap_dmtimer_init parameter type to usage
The clk_speed argument is used as input to clocksource_hz2mult() which
has a u32 as parameter. While u32 and unsigned are equivalent in
barebox, it's still a good idea to differentiate between "normal"-sized
integers and integers that are explicitly meant to be exactly 32-bit.

No functional change.

Suggested-by: Daniel Lezcano <daniel.lezcano@linaro.org>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250429132550.3768805-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>