firmware: ti_sci: handle Asel
This adds support for the Asel aka "Address selection" bits to ti_sci.

These bits become important when the DDR firewalls are enabled. With DDR
firewalls the buffer addresses used by the PKTDMA engine need to be
tagged with the correct Asel value, otherwise they will be blocked by
the firewall.

Link: https://lore.barebox.org/20250429-k3-asel-v1-3-6f39afaa2dfe@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>