dma: k3-udma: Handle Asel
This adds support for the Asel aka "Address selection" bits to the
k3-udma driver.

These bits become important when the DDR firewalls are enabled. With DDR
firewalls the buffer addresses used by the PKTDMA engine need to be
tagged with the correct Asel value, otherwise they will be blocked by
the firewall.

The Asel values are read from the device tree and need to be placed in
bits 48-51 of the physical DMA addresses.

Link: https://lore.barebox.org/20250429-k3-asel-v1-4-6f39afaa2dfe@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>