fb: generate mode name if missing
If a mode name is not set, the device parameters will format the NULL
mode name as <NULL>. Instead, let's just generate a name and use that.

If multiple modes have the same resolution and refresh rate, we will end
up with duplicate names, but that's ok as this is just a fallback.
Drivers can always assign proper names to their modes and most do.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250506062117.3090793-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>