driver: base: invert driver match callback for Linux compatibility
Both Linux and barebox have a struct bus_type::match callback that
checks whether a driver matches a device.

In Linux, a match is indicated by a non-zero return code, but in barebox
zero is used for that purpose.

To avoid this annoying pitfall, let's switch over barebox to the Linux
scheme. While at it, we also adapt the prototype to make the driver
const as is the case with Linux. This gives us some extra certainty that
all bus types have indeed been adapted.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250506063446.3133582-8-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>