mci: atmel_mci: switch PBL implementation to common mci-pbl support
We don't reinitialize the SD-Card when barebox is used as a first stage,
because it's already initialized by the BootROM and thus we don't use
the MMC stack. Recently, a very stripped down version was added that
doesn't initialize cards, but determines wheter a card is high or
standard capacity and then starts reading blocks directly.
Make use of it.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250506115652.1134337-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>