clocksource: ti-dm: fix EPROBE_DEFER handling
barebox initializes only a single ti-dm timer and stores its non-NULL
MMIO base address to the global base variable to indicate a successful
probe.

If clock_get were to fail with any error code, base would remain set,
hindering future probes of the driver. This would be especially
bothersome with EPROBE_DEFER if deep probe is not enabled.

Fix this by removing the early setting of the base variable.

Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>
Link: https://lore.barebox.org/20250519185917.459180-1-a.fatoum@barebox.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>