ARM: omap: fix cpu_is_omap on CONFIG_ARM_MULTIARCH configurations
Normally, <mach/omap/generic.h> would only be included if we are running
on some OMAP SoC. This doesn't hold true anymore with CONFIG_ARM_MULTIARCH
or CONFIG_BOARD_ARM_GENERIC_DT, because we could have CONFIG_ARCH_AM33XX
enabled, but running on some other SoC, which the header doesn't know about.

This cpu_type scheme is only used on one other SoC and that's the i.MX
and there, there is a final #ifdef clause to fix this, so follow suit.

Reported-by: Roland Hieber <r.hieber@pengutronix.de>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Tested-by: Roland Hieber <rhi@pengutronix.de>
Link: https://lore.barebox.org/20250520121738.42686-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>