uncompress: fix prebuffering
When uncompressing from a fill function we prebuffer the first 32 bytes
in order to detect the filetype. This prebuffer is then fed back into
the input stream in uncompress_fill().

This currently fails when uncompress_fill() crosses the prebuffered
size and continues reading from the original fill function. Advance
the buffer in this case to append the data to the buffer instead of
overwriting the data we already copied from the prebuffer.

Link: https://lore.barebox.org/20250522-uncompress-fixes-v1-2-18e796fa5e45@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>