resource: record EFI memory type and attributes
While struct resource has some defines for attributes like
IORESOURCE_MEM_CACHEABLE, there are not enough bits to cover all
memory attributes and types specified by EFI.

As the upper 32-bit of the existing struct resource::flags member on
64-bit systems is unused, let's repurpose it to hold the memory type and
attributes. As we want to make use of this information even without EFI,
we add generic sounding constants that happen to be identical to the EFI
constant values.

Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>
Link: https://lore.barebox.org/20250526141914.185212-3-a.fatoum@barebox.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>