efi: use size_t for UINTN array sizes instead of unsigned long
UINTN is the native register wide integer type. We don't have a specific
typedef for this in barebox, instead using size_t in places and unsigned
long in others.

The type confusion is no problem on 64-bit platforms, where size_t is
just a typedef for long. On 32-bit platforms, size_t is a typedef for
int, which while having the same size as long, is a different type,
which leads to compiler warning and errors regarding use of different
types for pointer arguments or in printf format strings.

The nice way out would have just define the size_t type as alias for
unsigned long everywhere, but this leads to false positive printf format
string warnings, so let's clean up the EFI UINTN type usage:

  - Everywhere, where the UINTN is used for an array size or buffer
    length, use size_t

  - Everywhere else keep it as unsigned long. In future, these types may
    be replaces by more specific typedefs.

Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>
Link: https://lore.barebox.org/20250527212300.575031-12-a.fatoum@barebox.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>