clocksource: efi: use DIV_ROUND_DOWN_ULL for 64-bit devision
The denominator freq is 64-bit, but this worked out ok so far, because
the driver was only used on 64-bit systems. Now that we want to start
compiling it for 32-bit as well, make the division work there as well.

Note that nb_100ns is a valid result and will instruct the EFI firmware
to expire the timer at the next tick.

Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>
Link: https://lore.barebox.org/20250527212300.575031-15-a.fatoum@barebox.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>