string: add delimiter output parameter to strsep_unescaped
strsep overwrites the found delimiter with '\0' making it cumbersome for
callers that match on multiple delimiters to find out, which delimiter
actually matched.

Parsers that split on multiple delimiters are likely to want support
escaping them too, so let's add an extra output parameter to
strsep_unescaped to make it possible to retrieve the delimiter that was
ultimately overwritten.

The intention behind this change is to allow retrofitting existing
space-separated strings with a colon separator that has special
semantics for zero-size strings:

  "ayy  cee" -> old behavior: multiple spaces are concatenated
  "ayy::cee" -> empty string expands to some default value

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250528055814.1368888-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>