ARM: add CONFIG_SCRATCH_SIZE
CONFIG_SCRATCH_SIZE allows to configure the size of the scratch area for
passing data from the PBL to barebox proper.

It may be also used to pass data from the PBL to other firmware or use
it for passing data back to barebox, the size should be configurable.
Since this may be more than 32K, the size should be configurable.

Since other sizes in the memory layout are configurable as well, the
memory layout isn't static.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>
Reviewed-by: Marco Felsch <m.felsch@pengutronix.de>
Link: https://lore.barebox.org/20250528-rk3588-optee-v2-6-63070238dd13@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>