ARM: rockchip: pass device tree to TF-A
The upstream OP-TEE expects a device tree to be able to initialize the
dynamic shared memory. Therefore, barebox should pass a device tree that
contains memory nodes through the TF-A to OP-TEE.

OP-TEE may modify the passed fdt. Thus, barebox copies the fdt from the
fixed size rodata area to the rk_scratch area and allocates a
configurable memory area for the fdt to allow fdt modification.

OP-TEE has the CFG_DTB_MAX_SIZE config option for the maximum size of
the fdt. barebox must reserve at least that much memory for the fdt to
avoid out of bounds accesses from OP-TEE.

Unfortunately, the downstream TF-A fails to start if barebox passes its
device tree and it is not possible to detect if the loaded TF-A is able
to handle the device tree. Add a config option to pass the device tree
if it is possible.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>
Reviewed-by: Marco Felsch <m.felsch@pengutronix.de>
Link: https://lore.barebox.org/20250528-rk3588-optee-v2-8-63070238dd13@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>