scripts: add bareboxtlv host/target tool
In its simplest form, a barebox TLV has a 12-byte header, a sequence of
TLV entries, where each of tags and length are 2 bytes each (in big
endian), followed by a CRC.

The TLVs are easily generated. Add a tool to allow adding, removing and
verifying the header. A full fledged python-based tool is added in the
follow up commit.

Example uses:

  hex2bin tlvs | ./scripts/bareboxtlv -m 0xcfe01001 >tlv.blob
  ./scripts/bareboxtlv -m 0xcfe01001 -v tlv.blob && echo CRC correct

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250528143043.434446-7-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>