fbconsole: use only one default color
The default framebuffer color is (dark) white, which we also revert to
when resetting the default color and keeping style intact.

We also support the ANSI "reset all colors/styles" sequence and include
it in the default barebox prompt, but that one doesn't currently default
to white, but bright black (gray) instead.

Given that we have a pitch black background by default, dark white is
easier on the eyes than gray, so let's make that the singular default.

Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>
Link: https://lore.barebox.org/20250601210002.562733-4-a.fatoum@barebox.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>