fbconsole: reset cursor to origin when changing fonts
If we switch from one font to another, we update priv->rows and
priv->cols, but priv->x and priv->y remain at their old values.

This becomes problematic, when a bigger font is chosen, because the
coordinates may now exceed the maximum number of rows and lines,
triggering an out-of-bounds memory access, when they are scaled
according to the new font.

Fix this by resetting to (0, 0) whenever the number of columns and rows
changes. The end result looks neater anyway.

Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>
Link: https://lore.barebox.org/20250601210002.562733-5-a.fatoum@barebox.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>