logo: spell out CONFIG_ options for fixdep
To avoid rebuilding nearly the whole project when a header changes,
fixdep collects referenced CONFIG_ symbols in every file and prunes the
dependency list accordingly. This only works when we don't hide CONFIG_
symbols with preprocessor text substitution shenanigans.

logo.c does just that triggering a compile failure, when logos are
removed in config without touching the file. Spell out the config
symbols to fix this.

This is the only instance currently of /CONFIG_.*##/, so I believe we
don't have this bug elsewhere.

Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>
Link: https://lore.barebox.org/20250602060245.746307-1-a.fatoum@barebox.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>