video: give struct fb_videomode::pixclock a strong picoseconds_t type
To avoid confusion between struct fb_videomode::pixclock, which in
picoseconds and other timing structs, which are in
struct videomode::pixelclock, which is in Hz, let's make it an error to
assign one to the other. This should nudge video driver ports into the
right direction in future.

No functional change intended.

Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>
Link: https://lore.barebox.org/20250602061610.752815-2-a.fatoum@barebox.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>