treewide: dts: describe deep probe support in device tree
Now that we have CONFIG_DEEP_PROBE_DEFAULT, annotate our device trees to
explicitly opt-out and opt-in from deep probe.

The opt-out is necessary, so boards aren't broken by whether we have
CONFIG_DEEP_PROBE_DEFAULT enabled or not. The opt-in is optional,
because all deep-probe enabled boards have a BAREBOX_DEEP_PROBE_ENABLE
in board code, but in future we could omit that.

Reviewed-by: Marco Felsch <m.felsch@pengutronix.de>
Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>
Link: https://lore.barebox.org/20250602095738.822334-3-a.fatoum@barebox.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>