treewide: use only string literals as format string
With the addition of the __printf attribute to dev_set_name, we get a
few warnings now about use of variables instead of string literals.

This can be problematic if said variables were to contain format
specifiers, which would lead to memory safety issues.

This all these trivial cases by adding a "%s" as format string.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250605112533.1960069-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>