scripts/k3img: make more flexible
We had a template for the certificate in which each component has a
certain name and the file to use for each component was specified using
different options (--sysfw, --sysfwdata, --dmdata, --sbl and
--innerdata). All components have the same format and the name doesn't
matter. Instead of supporting different options just take a list of
components as input of the tool. Each component can be specified like:

filename:compType:bootCore:compOpts:destAddr

This not only makes the tool easier to follow, but also lets us specify
the compType, bootCore and compOpts options which need different values
for different SoCs.

While at it add support for the --help option.

Link: https://lore.barebox.org/20250605-arm-k3-am62l-v2-1-53257d4b2dd2@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>