ARM: dts: am62l: Fix assigned-clock-parents
The AM62l device trees we imported to barebox have problems with the
clock parents which is described in [1] and [2].

There is no way the Kernel or barebox SCMI clk drivers can work with the
current way the clk parents are specified in the downstream TF-A. This
problem is acknowledged by TI and will hopefully be fixed.

This patch fixes the clk parents the way that I hope TI will do it as
well. Currently this only works with a patched downstream TF-A.

[1] https://lore.kernel.org/all/aCcSG5ah12N0yOwi@pengutronix.de/
[2] https://review.trustedfirmware.org/c/TF-A/trusted-firmware-a/+/34834/1/plat/ti/k3/common/drivers/scmi/scmi_clock.c#165

Link: https://lore.barebox.org/20250605-arm-k3-am62l-v2-5-53257d4b2dd2@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>