ARM: Makefile: descend into mach-* for cleaning
When you run "make clean" for arm, it never visits mach-* directories
because machine-y is just empty. With this files generated with extra-y
like the .k3cfg files are not cleaned.

When cleaning, all machine directories are accumulated to machine-, so
let's pass them to common- to clean up those directories.

Link: https://lore.barebox.org/20250605-arm-k3-am62l-v2-7-53257d4b2dd2@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>