firmware: add ti-linux-firmware submodule
For K3 SoCs we need several binaries from the ti-linux-firmware
repository and currently we have to copy the needed files over from the
repository to firmware/ in barebox. To simplify this expect the
ti-linux-firmware repository to be checked out or linked at
firmware/ti-linux-firmware. For further convenience create a git
submodule for it.

Link: https://lore.barebox.org/20250605-arm-k3-am62l-v2-9-53257d4b2dd2@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>